pyproject.toml
[tool.poetry]
name = "application"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
plaid = "^0.1.7"
python-dotenv = "^0.19"
plaid-python = "^24.0.0"
flask-cors = "^4.0.1"
flask-wtf = "^1.2.1"
email-validator = "^2.2.0"
dash = "^2.17.1"
pandas = "^2.2.2"
flask-sqlalchemy = "^3.1.1"
dash-bootstrap-components = "^1.6.0"
flask-bcrypt = "^1.0.1"
flask-login = "^0.6.3"
pillow = "^10.4.0"

[tool.poetry.scripts]
start = "src.app:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
